refactor command service to use one method per file
Updated the command service to have the same structure as all the other services within platform/...
