+++ platform/command/service.go
@@ -2,11 +2,8 @@
 package command
 
 import (
-	"time"
-
-	"github.com/go-kit/kit/log"
-	"github.com/go-kit/kit/metrics"
 	"github.com/micromdm/micromdm/mdm/mdm"
+	"github.com/micromdm/micromdm/platform/pubsub"
 	"golang.org/x/net/context"
 )
 
@@ -14,54 +11,13 @@ type Service interface {
 	NewCommand(context.Context, *mdm.CommandRequest) (*mdm.CommandPayload, error)
 }
 
-// Middleware describes a service (as opposed to endpoint) middleware.
-type Middleware func(Service) Service
+type CommandService struct {
+	publisher pubsub.Publisher
+}
 
-// ServiceLoggingMiddleware returns a service middleware that logs the
-// parameters and result of each method invocation.
-func ServiceLoggingMiddleware(logger log.Logger) Middleware {
-	return func(next Service) Service {
-		return serviceLoggingMiddleware{
-			logger: logger,
-			next:   next,
-		}
+func New(pub pubsub.Publisher) (*CommandService, error) {
+	svc := CommandService{
+		publisher: pub,
 	}
-}
-
-func (mw serviceLoggingMiddleware) NewCommand(ctx context.Context, req *mdm.CommandRequest) (p *mdm.CommandPayload, err error) {
-	defer func(begin time.Time) {
-		mw.logger.Log(
-			"method", "NewCommand",
-			"error", err,
-			"took", time.Since(begin),
-		)
-	}(time.Now())
-	return mw.next.NewCommand(ctx, req)
-}
-
-type serviceLoggingMiddleware struct {
-	logger log.Logger
-	next   Service
-}
-
-// ServiceInstrumentingMiddleware returns a service middleware that tracks the
-// number of payloads created by the service.
-func ServiceInstrumentingMiddleware(p metrics.Counter) Middleware {
-	return func(next Service) Service {
-		return serviceInstrumentingMiddleware{
-			payloads: p,
-			next:     next,
-		}
-	}
-}
-
-type serviceInstrumentingMiddleware struct {
-	payloads metrics.Counter
-	next     Service
-}
-
-func (mw serviceInstrumentingMiddleware) NewCommand(ctx context.Context, req *mdm.CommandRequest) (*mdm.CommandPayload, error) {
-	p, err := mw.next.NewCommand(ctx, req)
-	mw.payloads.Add(1)
-	return p, err
+	return &svc, nil
 }
